Chivalry2.h UE4.h, remove unused
