+++ UnchainedPlugin/UnchainedPlugin.vcxproj.filters
@@ -48,7 +48,13 @@
     <ClInclude Include="include\sigs.h">
       <Filter>Source Headers</Filter>
     </ClInclude>
-    <ClInclude Include="Lib\MinHook\src\buffer.h">
+	<ClInclude Include="include\UE4.h">
+		<Filter>Source Headers</Filter>
+	</ClInclude>
+	  <ClInclude Include="include\Chivalry2.h">
+		  <Filter>Source Headers</Filter>
+	  </ClInclude>
+	<ClInclude Include="Lib\MinHook\src\buffer.h">
       <Filter>Lib Headers</Filter>
     </ClInclude>
     <ClInclude Include="Lib\MinHook\src\hde\hde32.h">
