+++ UnchainedPlugin/include/UE4.h
@@ -1,0 +1,16 @@
+#pragma once
+#include <Windows.h>
+
+// UE Types
+
+struct FString {
+	FString(const wchar_t* str) {
+		this->letter_count = lstrlenW(str) + 1;
+		this->max_letters = this->letter_count;
+		this->str = const_cast<wchar_t*>(str);
+	}
+
+	wchar_t* str;
+	int letter_count;
+	int max_letters;
+};
